Refactor avatar and game
