+++ src/main/java/seedu/duke/game/achievement/AddTask.java
@@ -1,0 +1,7 @@
+package seedu.duke.game.achievement;
+
+public class AddTask extends Achievements {
+
+
+
+}
