+++ src/main/java/seedu/duke/game/achievement/ConsecutiveLogin.java
@@ -1,0 +1,8 @@
+package seedu.duke.game.achievement;
+
+public class ConsecutiveLogin extends Achievements {
+
+
+
+
+}
