+++ src/main/java/seedu/duke/game/achievement/DoneTask.java
@@ -1,0 +1,6 @@
+package seedu.duke.game.achievement;
+
+import seedu.duke.game.achievement.Achievements;
+
+public class DoneTask extends Achievements {
+}
