+++ src/test/java/seedu/duke/game/avatar/Avatar.java
@@ -1,12 +1,12 @@
-package seedu.duke.avatar;
+package seedu.duke.game.avatar;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- *  Test class for Level.
+ *  Test class for Avatar.
  */
-public class Level {
+public class Avatar {
 
     @Test
     /**
